feat login logout in frontend and fix some logic in code fe and header
