+++ FrontEnd/model-trade-app/app/components/layout/navigation/Header.tsx
@@ -1,192 +1,3 @@
-// // components/layout/Header.tsx
-// "use client";
-// import { BsChatDots } from "react-icons/bs";
-// import { FaRegUser } from "react-icons/fa";
-// import { MdOutlineShoppingCart } from "react-icons/md";
-
-// import { useState, useEffect } from "react";
-// import Link from "next/link";
-// import Image from "next/image";
-// import CategoryMenu from "./CategoryMenu";
-// import SearchBar from "./Search";
-
-// interface HeaderProps {
-//   openModal?: () => void;
-// }
-
-// const Header: React.FC<HeaderProps> = ({ openModal }) => {
-//   const [searchQuery, setSearchQuery] = useState("");
-//   const [isNavVisible, setIsNavVisible] = useState(true);
-//   const [lastScrollY, setLastScrollY] = useState(0);
-
-//   // Xử lý sự kiện cuộn
-//   const controlNavbar = () => {
-//     if (typeof window !== "undefined") {
-//       // Nếu cuộn xuống và đã cuộn xuống ít nhất 20px, ẩn navbar
-//       if (window.scrollY > lastScrollY && window.scrollY > 20) {
-//         setIsNavVisible(false);
-//       }
-//       // Nếu cuộn lên hoặc ở đầu trang, hiện navbar
-//       else {
-//         setIsNavVisible(true);
-//       }
-
-//       // Cập nhật vị trí cuộn cuối cùng
-//       setLastScrollY(window.scrollY);
-//     }
-//   };
-//   // Thêm event listener khi component mount
-//   useEffect(() => {
-//     if (typeof window !== "undefined") {
-//       window.addEventListener("scroll", controlNavbar);
-
-//       // Cleanup event listener khi component unmount
-//       return () => {
-//         window.removeEventListener("scroll", controlNavbar);
-//       };
-//     }
-//   }, [lastScrollY]);
-
-//   const handleSearch = (e: React.FormEvent) => {
-//     e.preventDefault();
-//     // Xử lý tìm kiếm
-//     console.log("Searching for:", searchQuery);
-//   };
-
-//   return (
-//     <header className=" border-gray-200 w-full top-0 left-0  sticky z-10  ">
-//       {/* Top Bar */}
-//       <div className="bg-primary py-2 relative lg:h-17   z-20">
-//         <div className="container mx-auto flex justify-between items-center px-2 md:px-1">
-//           {/* Logo */}
-//           <Link
-//             href="/home"
-//             className="top-0 left-0 flex items-start ml-4 md:mx-8 "
-//           >
-//             <div className="relative  pt-2 lg:pt-0">
-//               <Image
-//                 // src="https://static.chotot.com/storage/APP_WRAPPER/logo/chotot-logo-appwrapper.png"
-//                 src="https://res.cloudinary.com/dxgqzu2vb/image/upload/v1748609195/nludev/product/Efigure.png_20250530194627.png"
-//                 alt="Logo"
-//                 width={120}
-//                 height={50}
-//                 className="  "
-//               />
-//             </div>
-//           </Link>
-//           <div className="hidden md:block w-full">
-//             <SearchBar />
-//           </div>
-
-//           <div className="flex items-center justify-end ">
-//             {/* Login Link */}
-//             <Link href="/login" className="flex items-center mx-3">
-//               <div className="flex w-fit items-center ">
-//                 <FaRegUser className="text-2xl mr-2" />
-//                 <span className="flex text-base items-center">Login</span>
-//               </div>
-//             </Link>
-
-//             {/* Cart Link */}
-//             <Link href="/cart" className=" flex items-center mx-3">
-//               <div className="flex w-fit items-center ">
-//                 <MdOutlineShoppingCart className="text-2xl mr-2" />
-//                 <p className="flex text-base items-center">Cart</p>
-//               </div>
-//             </Link>
-
-//             <Link href="/chat" className=" flex items-center mx-3">
-//               <div className="flex w-fit items-center ">
-//                 <BsChatDots className="text-2xl mr-2" />
-//                 <p className="flex text-base items-center">Chat</p>
-//               </div>
-//             </Link>
-
-//             <button
-//               className=" cursor-pointer flex items-center w-full mx-3"
-//               onClick={openModal}
-//             >
-//               <div className="flex h-auto w-fit items-center justify-center  ">
-//                 <BsChatDots className="text-2xl mr-2" />
-//                 <p className="flex  whitespace-nowrap text-base">Bài viết</p>
-//               </div>
-//             </button>
-
-//             <Link className=" flex items-center mx-3" href="/notification">
-//               <div className=" flex h-auto w-fit items-center">
-//                 <BsChatDots className="text-2xl mr-2" />
-//                 <p className="flex text-base  whitespace-nowrap">Thông báo</p>
-//               </div>
-//             </Link>
-//           </div>
-//         </div>
-//       </div>
-
-//       {/* Navigation Bar */}
-//       {/* <nav className="bg-blue-700 text-white transition-transform duration-300 ${
-//           isNavVisible ? 'transform-none' : '-translate-y-full'
-//         "> */}
-//       <nav
-//         className={`bg-primary text-white transition-transform duration-300 ease-in-out border-none relative z-10 ${
-//           isNavVisible ? "translate-y-0" : "-translate-y-full"
-//         }`}
-//       >
-//         <div className=" container h-12 md:h-8  flex  md:items-center  ">
-//           <CategoryMenu />
-//           <div className=" hidden container  md:mx-auto px-2 md:flex justify-items-start items-start md:space-x-6 ">
-//             <Link
-//               href="/newsletter"
-//               className=" text-sm font-medium whitespace-nowrap "
-//             >
-//               Join Our Newsletter
-//             </Link>
-//             <Link
-//               href="/how-to"
-//               className="text-sm font-medium whitespace-nowrap"
-//             >
-//               How to Use It
-//             </Link>
-//             <Link
-//               href="/warehouse"
-//               className=" text-sm font-medium whitespace-nowrap"
-//             >
-//               What's Private Warehouse?
-//             </Link>
-//             <Link
-//               href="/help"
-//               className=" text-sm font-medium whitespace-nowrap"
-//             >
-//               Help Center
-//             </Link>
-//             <Link
-//               href="/wishlist"
-//               className=" text-sm font-medium whitespace-nowrap"
-//             >
-//               Wish List
-//               <svg
-//                 xmlns="http://www.w3.org/2000/svg"
-//                 className="h-4 w-4 inline-block ml-1"
-//                 fill="none"
-//                 viewBox="0 0 24 24"
-//                 stroke="currentColor"
-//               >
-//                 <path
-//                   strokeLinecap="round"
-//                   strokeLinejoin="round"
-//                   strokeWidth={2}
-//                   d="M19 9l-7 7-7-7"
-//                 />
-//               </svg>
-//             </Link>
-//           </div>
-//         </div>
-//       </nav>
-//     </header>
-//   );
-// };
-
-// export default Header;
-// // components/layout/Header.tsx
 "use client";
 import { BsChatDots } from "react-icons/bs";
 import { FaRegUser } from "react-icons/fa";
@@ -196,6 +7,8 @@ import Link from "next/link";
 import Image from "next/image";
 import CategoryMenu from "./CategoryMenu";
 import SearchBar from "./Search";
+import { useRouter } from "next/navigation";
+import { IoLogOutOutline } from "react-icons/io5";
 
 interface HeaderProps {
   openModal?: () => void;
@@ -206,11 +19,12 @@ const Header: React.FC<HeaderProps> = ({ openModal }) => {
   const [isNavVisible, setIsNavVisible] = useState(true);
   const [lastScrollY, setLastScrollY] = useState(0);
   const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const router = useRouter();
 
   // Kiểm tra trạng thái đăng nhập dựa trên token
   useEffect(() => {
     const token = localStorage.getItem("token");
-    setIsAuthenticated(!!token); // Đặt true nếu token tồn tại
+    setIsAuthenticated(!!token);
   }, []);
 
   // Xử lý sự kiện cuộn
@@ -237,36 +51,43 @@ const Header: React.FC<HeaderProps> = ({ openModal }) => {
     console.log("Searching for:", searchQuery);
   };
 
+  // Xử lý đăng xuất
+  const handleLogout = () => {
+    localStorage.removeItem("token"); // Xóa token khỏi localStorage
+    setIsAuthenticated(false); // Cập nhật trạng thái đăng nhập
+    router.push("/login"); // Chuyển hướng về trang đăng nhập
+  };
+
+  //  Xử lý nhấp vào logo reload trang
+  const handleLogoClick = (e: React.MouseEvent<HTMLAnchorElement>) => {
+    e.preventDefault(); // Ngăn chặn hành vi mặc định của thẻ a
+    window.location.reload(); // Tải lại trang
+  };
+
   return (
     <header className="border-gray-200 w-full top-0 left-0 sticky z-10">
       <div className="bg-primary py-2 relative lg:h-17 z-20">
-        <div className="container mx-auto flex justify-between items-center px-2 md:px-1">
+        <div className="container mx-auto flex justify-between items-center ">
           <Link
             href="/home"
             className="top-0 left-0 flex items-start ml-4 md:mx-8"
+            onClick={handleLogoClick} // Thêm sự kiện nhấp vào logo
           >
-            <div className="relative pt-2 lg:pt-0">
+            <div className="relative  lg:pt-0">
               <Image
                 src="https://res.cloudinary.com/dxgqzu2vb/image/upload/v1748609195/nludev/product/Efigure.png_20250530194627.png"
                 alt="Logo"
-                width={120}
-                height={50}
+                width={60}
+                height={30}
                 className=""
               />
             </div>
           </Link>
-          <div className="hidden md:block w-full">
+          <div className="hidden md:block w-3/5">
             <SearchBar />
           </div>
           <div className="flex items-center justify-end">
-            <Link href="/login" className="flex items-center mx-3">
-              <div className="flex w-fit items-center">
-                <FaRegUser className="text-2xl mr-2" />
-                <span className="flex text-base items-center">Login</span>
-              </div>
-            </Link>
-
-            {/* Hiển thị các thẻ chỉ khi đã đăng nhập */}
+            {/* Hiển thị các thẻ và nút đăng xuất chỉ khi đã đăng nhập */}
             {isAuthenticated && (
               <>
                 <Link href="/cart" className="flex items-center mx-3">
@@ -301,7 +122,24 @@ const Header: React.FC<HeaderProps> = ({ openModal }) => {
                     </p>
                   </div>
                 </Link>
+
+                <button
+                  onClick={handleLogout}
+                  className="flex items-center mx-3 hover:text-red-700"
+                >
+                  <IoLogOutOutline className="text-2xl mr-2" />
+                  <span className="flex text-base items-center">Logout</span>
+                </button>
               </>
+            )}
+            {/* Hiển thị nút đăng nhập nếu chưa đăng nhập */}
+            {!isAuthenticated && (
+              <Link href="/login" className="flex items-center mx-3 mr-5">
+                <div className="flex w-fit items-center">
+                  <FaRegUser className="text-2xl mr-2" />
+                  <span className="flex text-base items-center">Login</span>
+                </div>
+              </Link>
             )}
           </div>
         </div>
